fixed small error

+++ lib/models/group.dart
@@ -4,7 +4,7 @@ import 'package:cloud_firestore/cloud_firestore.dart';
 
 class Group {
   String name;
-  Float total;
+  double total;
   Timestamp? createdOn;
   List<String>? members;
 
@@ -14,7 +14,7 @@ class Group {
   Group.fromJson(Map<String, Object?> json)
       : this(
           name: json['name']! as String,
-          total: json['total']! as Float,
+          total: json['total']! as double,
           createdOn: json['createdOn']! as Timestamp,
           members: (json['members'] as List<dynamic>)
               .map((e) => e as String)
@@ -23,7 +23,7 @@ class Group {
 
   Group copyWith(
       {String? name,
-      Float? total,
+      double? total,
       Timestamp? createdOn,
       List<String>? members}) {
     return Group(
